remove the domains option from GoogleAppsGateway
We were using these to retrieve the users for multiple domains that are
active in a google apps account, but it turns out we can achieve the
same result in a single request with the 'customer: "my_customer"'
API param.

I've also bumped the number of users and groups that can be returned to
500. That makes this plugin only useful for accounts with 500 or fewer
users, but that's no worse than the current situation.