reimplement GoogleAppsGateway#two_factor_users with standard users.list API
Until now we've been using the reporting API to determine if a user has
two factor auth enabled, but it seems like that data is now available
via the standard users.list API.

Standardising on users.list is advantageous because it reduces the
number of oauth scoeps we need permission for, and changes all the user
related methods of GoogleAppsGateway to return GoogleUser value objects.

While I was there, I also implemented GoogleAppsGateway#super_admins and
GoogleAppsGateway#delegated_admin because they were easy and were
useful.